fix(blocks): Fix path stairs being un-climbable
Path stairs could not be climbed correctly as the player was pushed off them. This was caused by an incorrect collision check.

Adding `.suffocates(Blocks::never)` resolves the issue by disabling the suffocation check for this block type. This fix was the result of a two-day debugging session.

As a minor change, the `ModelHelper` was also moved into the `helper` package.